Re-quantize and scale preson detection scores

+++ examples/person_detection/main/detection_responder.h
@@ -29,6 +29,6 @@ limitations under the License.
 // image is considered to contain a person.  This threshold may be adjusted for
 // particular applications.
 void RespondToDetection(tflite::ErrorReporter* error_reporter,
-                        int8_t person_score, int8_t no_person_score);
+                        float person_score, float no_person_score);
 
 #endif  // TENSORFLOW_LITE_MICRO_EXAMPLES_PERSON_DETECTION_DETECTION_RESPONDER_H_
